Fix output dirs

+++ modules/extracting_IMD_matrices.nf
@@ -15,7 +15,15 @@ process 'extracting_matrices' {
    script:
    """
    awk -v RS= '{print > ("${matrices}."NR".txt")}' ${matrices}
-
-   mv ${id}*matrices.1.txt ${id}_${params.align}_3d_${params.trimmer}.matrix
+   if [ -f *columns*.txt ]; then
+      mv ${id}*matrices.1.txt ${id}_${params.align}_3d_${params.trimmer}.matrix
+   else 
+      for file in *matrices.1.txt; do
+         new_file="\${file%.txt}.matrix"
+         # Rename the file
+         mv \${file} \${new_file}
+         echo "File \${file} renamed to \${new_file}" 
+      done
+   fi
    """
 }
